Application/requirements.txt
deep_translator==1.11.4
langchain_community==0.3.20
langchain_core==0.3.49
langchain_openai==0.3.11
langdetect==1.0.9
langgraph==0.3.21
langsmith==0.3.19
mysql-connector-python==9.2.0
pandas==2.2.3
python-dotenv==1.1.0
redis==5.2.1
scikit_learn==1.6.1
streamlit==1.43.2
duckduckgo_search==7.5.5